Base upstream 2.9.10 source

+++ tensorflow/python/ops/ragged/ragged_print_op_test.py
@@ -20,12 +20,9 @@ ragged_dispatch.py.
 
 import os.path
 import tempfile
-
 from absl.testing import parameterized
-
 from tensorflow.python.eager import def_function
 from tensorflow.python.framework import constant_op
-from tensorflow.python.framework import dtypes
 from tensorflow.python.framework import test_util
 from tensorflow.python.ops import logging_ops
 from tensorflow.python.ops import sparse_ops
@@ -179,9 +176,8 @@ class RaggedToStringTest(test_util.TensorFlowTestCase, parameterized.TestCase):
 
   def testRaggedToStringUnknownRank(self):
 
-    @def_function.function(input_signature=[
-        ragged_tensor.RaggedTensorSpec(ragged_rank=1, dtype=dtypes.int32)
-    ])
+    @def_function.function(
+        input_signature=[ragged_tensor.RaggedTensorSpec(ragged_rank=1)])
     def f(rt):
       return ragged_string_ops.ragged_tensor_to_string(rt)
